pyproject.toml
[tool.poetry]
name = "fact-sphere"
description = "A library for Portal 2 Fact Sphere facts."
version = "1.0.0"

license = "MIT"

authors = ["thebigmunch <mail@thebigmunch.me>"]

readme = "README.md"

repository = "https://github.com/thebigmunch/fact-sphere"
homepage = "https://github.com/thebigmunch/fact-sphere"

[tool.poetry.dependencies]
python = "^3.6"

attrs = "^18.1"

flake8 = { version = "^3.5", optional = true }
flake8-builtins = { version = "^1.0", optional = true }
flake8-import-order = { version = "^0.18", optional = true }
flake8-import-order-tbm = { version = "^1.0.0", optional = true }
poetry = { version = "^0.9", optional = true }

[tool.poetry.extras]
dev = [ "flake8", "flake8-builtins", "flake8-import-order", "flake8-import-order-tbm", "poetry" ]
lint = [ "flake8", "flake8-builtins", "flake8-import-order", "flake8-import-order-tbm" ]
